refactor(stm32mp1): update tamp_bkpr return type
tamp_bkpr return a register address. So use uintptr_t instead of uin32_t

Reviewed-on: https://gerrit.st.com/c/mpu/oe/st/tf-a/+/238700
Tested-by: Nicolas TOROMANOFF <nicolas.toromanoff@st.com>

+++ plat/st/stm32mp1/bl2_plat_setup.c
@@ -168,9 +168,9 @@ void bl2_platform_setup(void)
 
 	if (!stm32mp1_ddr_is_restored()) {
 #if STM32MP15
-		uint32_t bkpr_core1_magic =
+		uintptr_t bkpr_core1_magic =
 			tamp_bkpr(BOOT_API_CORE1_MAGIC_NUMBER_TAMP_BCK_REG_IDX);
-		uint32_t bkpr_core1_addr =
+		uintptr_t bkpr_core1_addr =
 			tamp_bkpr(BOOT_API_CORE1_BRANCH_ADDRESS_TAMP_BCK_REG_IDX);
 
 		/* Clear backup register */
