+++ plat/st/stm32mp1/sp_min/sp_min_setup.c
@@ -230,10 +230,10 @@ void sp_min_plat_fiq_handler(uint32_t id)
  ******************************************************************************/
 static uintptr_t get_saved_pc(void)
 {
-	uint32_t bkpr_core1_addr =
+	uintptr_t bkpr_core1_addr =
 		tamp_bkpr(BOOT_API_CORE1_BRANCH_ADDRESS_TAMP_BCK_REG_IDX);
 	uint32_t saved_pc;
-	uint32_t bkpr_core1_magic =
+	uintptr_t bkpr_core1_magic =
 		tamp_bkpr(BOOT_API_CORE1_MAGIC_NUMBER_TAMP_BCK_REG_IDX);
 	uint32_t magic_nb;
